:truck: Move details tab components around again.
None of these components are generic, so not really sure where they should go, only that they should be in folders matching the parents they are used in.
